+++ src/types/user.rs
@@ -46,65 +46,73 @@ impl User {
     }
 
     /// Returns an URL that links to this user in the form of
-    /// `tg://user/?id=<...>`
+    /// `tg://user/?id=<...>`.
     pub fn url(&self) -> reqwest::Url {
-        reqwest::Url::parse(&format!("tg://user/?id={}", self.id)).unwrap()
+        self.id.url()
     }
 
-    /// Returns `true` if this is special user used by telegram bot API to
+    /// Returns an URL that links to this user in the form of `t.me/<...>`.
+    /// Returns `None` if `self.username.is_none()`.
+    pub fn tme_url(&self) -> Option<reqwest::Url> {
+        Some(
+            format!("https://t.me/{}", self.username.as_ref()?)
+                .parse()
+                .unwrap(),
+        )
+    }
+
+    /// Returns an URL that links to this user in the form of `t.me/<...>` or
+    /// `tg://user/?id=<...>`, preferring `t.me` one when possible.
+    pub fn preferably_tme_url(&self) -> reqwest::Url {
+        self.tme_url().unwrap_or_else(|| self.url())
+    }
+
+    /// Returns `true` if this is the special user used by telegram bot API to
     /// denote an anonymous user that sends messages on behalf of a group.
     pub fn is_anonymous(&self) -> bool {
-        // https://github.com/tdlib/td/blob/4791fb6a2af0257f6cad8396e10424a79ee5f768/td/telegram/ContactsManager.cpp#L4941-L4943
-        const ANON_ID: UserId = UserId(1087968824);
-
         // Sanity check
         debug_assert!(
-            (self.id != ANON_ID)
+            !self.id.is_anonymous()
                 || (self.is_bot
                     && self.first_name == "Group"
                     && self.last_name.is_none()
                     && self.username.as_deref() == Some("GroupAnonymousBot"))
         );
 
-        self.id == ANON_ID
+        self.id.is_anonymous()
     }
 
-    /// Returns `true` if this is special user used by telegram bot API to
+    /// Returns `true` if this is the special user used by telegram bot API to
     /// denote an anonymous user that sends messages on behalf of a channel.
     pub fn is_channel(&self) -> bool {
-        // https://github.com/tdlib/td/blob/4791fb6a2af0257f6cad8396e10424a79ee5f768/td/telegram/ContactsManager.cpp#L4945-L4947
-        const ANON_CHANNEL_ID: UserId = UserId(136817688);
-
         // Sanity check
         debug_assert!(
-            (self.id != ANON_CHANNEL_ID)
+            !self.id.is_channel()
                 || (self.is_bot
                     && self.first_name == "Group"
                     && self.last_name.is_none()
                     && self.username.as_deref() == Some("GroupAnonymousBot"))
         );
 
-        self.id == ANON_CHANNEL_ID
+        self.id.is_channel()
     }
 
-    /// Returns `true` if this is special user used by telegram itself.
+    /// Returns `true` if this is the special user used by telegram itself.
     ///
     /// It is sometimes also used as a fallback, for example when a channel post
     /// is automatically forwarded to a group, bots in a group will get a
     /// message where `from` is the Telegram user.
     pub fn is_telegram(&self) -> bool {
-        const TELEGRAM_USER_ID: UserId = UserId(777000);
-
         // Sanity check
         debug_assert!(
-            (self.id != TELEGRAM_USER_ID)
+            !self.id.is_telegram()
                 || (!self.is_bot
                     && self.first_name == "Telegram"
                     && self.last_name.is_none()
                     && self.username.is_none())
         );
 
-        self.id == TELEGRAM_USER_ID
+        self.id.is_telegram()
     }
 }
 
@@ -132,5 +140,47 @@ mod tests {
         };
         let actual = serde_json::from_str::<User>(json).unwrap();
         assert_eq!(actual, expected)
+    }
+
+    #[test]
+    fn convenience_methods_work() {
+        let user_a = User {
+            id: UserId(43),
+            is_bot: false,
+            first_name: "First".to_owned(),
+            last_name: Some("Last".to_owned()),
+            username: Some("aaaaaaaaaaaaaaaa".to_owned()),
+            language_code: None,
+        };
+
+        let user_b = User {
+            id: UserId(44),
+            is_bot: false,
+            first_name: ".".to_owned(),
+            last_name: None,
+            username: None,
+            language_code: None,
+        };
+
+        assert_eq!(user_a.full_name(), "First Last");
+        assert_eq!(user_b.full_name(), ".");
+
+        assert_eq!(user_a.mention(), Some("@aaaaaaaaaaaaaaaa".to_owned()));
+        assert_eq!(user_b.mention(), None);
+
+        assert_eq!(
+            user_a.tme_url(),
+            Some("https://t.me/aaaaaaaaaaaaaaaa".parse().unwrap())
+        );
+        assert_eq!(user_b.tme_url(), None);
+
+        assert_eq!(
+            user_a.preferably_tme_url(),
+            "https://t.me/aaaaaaaaaaaaaaaa".parse().unwrap()
+        );
+        assert_eq!(
+            user_b.preferably_tme_url(),
+            "tg://user/?id=44".parse().unwrap()
+        );
     }
 }