+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## unreleased
 
+### Added
+
+- `UserId::{url, is_anonymous, is_channel, is_telegram}` convenience functions ([#197][pr197]) 
+- `User::{tme_url, preferably_tme_url}` convenience functions ([#197][pr197])
+- `Me::username` and `Deref<Target = User>` implementation for `Me` ([#197][pr197])
+- `Me::{mention, tme_url}` ([#197][pr197])
+
+[pr197]: https://github.com/teloxide/teloxide-core/pull/197
+
 ### Changed
 
 - `user.id` now uses `UserId` type, `ChatId` now represents only _chat id_, not channel username, all `chat_id` function parameters now accept `Recipient` [**BC**]